+++ src/general/dto/parameter.dto.ts
@@ -44,6 +44,22 @@ export const SCHEMA_VALID_OUT_SRS: Readonly<SchemaObject> = {
   ],
 };
 
+export const SCHEMA_VALID_RETURN_BUFFER_GEOMETRY: Readonly<SchemaObject> = {
+  anyOf: [
+    {
+      properties: {
+        returnBufferGeometry: { enum: [false] },
+      },
+    },
+    {
+      required: ['buffer'],
+      properties: {
+        buffer: { not: { enum: [0] } },
+      },
+    },
+  ],
+};
+
 @ApiExtraModels(EsriJsonDto, GeoJSONFeatureDto, GeoJSONFeatureCollectionDto)
 export class ParameterDto implements GeospatialRequest {
   @ApiProperty({
@@ -125,8 +141,31 @@ export class ParameterDto implements GeospatialRequest {
   outSRS: number;
 }
 
-export class IntersectParameterDto extends ParameterDto {}
-export class NearestNeighbourParameterDto extends ParameterDto {
+export const MAX_BUFFER_DISTANCE = 20000;
+
+export class BufferParameterDto extends ParameterDto {
+  @ApiProperty({
+    required: false,
+    type: Number,
+    example: 100,
+    description: 'Buffer distance around each input geometry in meters',
+    minimum: 0,
+    maximum: MAX_BUFFER_DISTANCE,
+  })
+  buffer?: number;
+
+  @ApiProperty({
+    required: false,
+    default: false,
+    description:
+      'Return the calculated buffer geometry of the input geometry in the response',
+    example: false,
+  })
+  returnBufferGeometry?: boolean;
+}
+
+export class IntersectParameterDto extends BufferParameterDto {}
+export class NearestNeighbourParameterDto extends BufferParameterDto {
   @ApiProperty({ example: 3 })
   count: number;
 
@@ -140,4 +179,4 @@ export class ValuesAtPointParameterDto extends ParameterDto {
   })
   topics: string[];
 }
-export class WithinParameterDto extends ParameterDto {}
+export class WithinParameterDto extends BufferParameterDto {}